Projeto1/pyproject.toml
[project]
name = "introdeeplearning"
version = "0.1.0"
description = "YOLOv3 Object Detection with PyTorch"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "ultralytics>=8.0.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
